Tests/LibWeb/Text/expected/abspos-replaced-auto-insets-static-position.txt

top: 70px
right: 170px
bottom: 90px
left: 70px
offset within containing block: 70,50
